Modified the type to bigint for code_num field in codes table
The type was giving convertion problem when providing code with some number in it that could be above a bigint.
It caused problem at some lines to be imported.
Thanks to Franck Theeten for the solution ;)

fix #1794

+++ data/db/createtables.sql
@@ -1195,7 +1195,7 @@ create table codes
         full_code_indexed varchar not null,
         code_date timestamp not null default '0001-01-01 00:00:00',
         code_date_mask integer not null default 0,
-        code_num integer default 0,
+        code_num bigint default 0,
         constraint pk_codes primary key (id),
         constraint unq_codes unique (referenced_relation, record_id, full_code_indexed,code_category)
        )
